+++ packages/core/List/index.js
@@ -92,7 +92,7 @@ const mapNativeTypeToKeystonType = (type, listKey, fieldPath) => {
 };
 
 module.exports = class List {
-  constructor(key, config, { getListByKey, adapter, defaultAccess, getAuth }) {
+  constructor(key, config, { getListByKey, getGraphQLQuery, adapter, defaultAccess, getAuth }) {
     this.key = key;
 
     // 180814 JM TODO: Since there's no access control specified, this implicitly makes name, id or {labelField} readable by all (probably bad?)
@@ -192,6 +192,40 @@ module.exports = class List {
     this.views = mapKeys(sanitisedFieldsConfig, fieldConfig => ({
       ...fieldConfig.type.views,
     }));
+
+    this.hooksActions = {
+      /**
+       * @param queryString String A graphQL query string
+       * @param options.skipAccessControl Boolean By default access control _of
+       * the user making the initial request_ is still tested. Disable all
+       * Access Control checks with this flag
+       * @param options.variables Object The variables passed to the graphql
+       * query for the given queryString.
+       *
+       * @return Promise<Object> The graphql query response
+       */
+      query: context => (queryString, { skipAccessControl = false, variables } = {}) => {
+        let passThroughContext = context;
+
+        if (skipAccessControl) {
+          passThroughContext = {
+            ...context,
+            getListAccessControlForUser: () => true,
+            getFieldAccessControlForUser: () => true,
+          };
+        }
+
+        const graphQLQuery = getGraphQLQuery();
+
+        if (!graphQLQuery) {
+          return Promise.reject(
+            new Error('No executable schema is available. Have you setup `@voussoir/server`?')
+          );
+        }
+
+        return graphQLQuery(queryString, passThroughContext, variables);
+      },
+    };
   }
 
   getAdminMeta() {
@@ -471,23 +505,24 @@ module.exports = class List {
     return mutations;
   }
 
-  checkFieldAccess(operation, existingItem, data, context, { gqlName, extraData = {} }) {
+  checkFieldAccess(operation, itemsToUpdate, context, { gqlName, extraData = {} }) {
     const restrictedFields = [];
 
-    this.fields
-      .filter(field => field.path in data)
-      .forEach(field => {
-        const access = context.getFieldAccessControlForUser(
-          this.key,
-          field.path,
-          existingItem,
-          operation
-        );
-        if (!access) {
-          restrictedFields.push(field.path);
-        }
-      });
-
+    itemsToUpdate.forEach(({ existingItem, data }) => {
+      this.fields
+        .filter(field => field.path in data)
+        .forEach(field => {
+          const access = context.getFieldAccessControlForUser(
+            this.key,
+            field.path,
+            existingItem,
+            operation
+          );
+          if (!access) {
+            restrictedFields.push(field.path);
+          }
+        });
+    });
     if (restrictedFields.length) {
       this._throwAccessDenied(operation, context, gqlName, extraData, { restrictedFields });
     }
@@ -638,8 +673,8 @@ module.exports = class List {
 
         [this.gqlNames.listMetaName]: (_, args, context) => this.listMeta(context),
 
-        [this.gqlNames.itemQueryName]: (_, { where: { id } }, context) =>
-          this.itemQuery(id, context, this.gqlNames.itemQueryName),
+        [this.gqlNames.itemQueryName]: (_, args, context) =>
+          this.itemQuery(args, context, this.gqlNames.itemQueryName),
       };
     }
 
@@ -711,7 +746,12 @@ module.exports = class List {
     };
   }
 
-  async itemQuery(id, context, gqlName) {
+  async itemQuery(
+    // prettier-ignore
+    { where: { id } },
+    context,
+    gqlName
+  ) {
     const operation = 'read';
     graphqlLogger.debug({ id, operation, type: opToType[operation], gqlName }, 'Start query');
 
@@ -728,7 +768,11 @@ module.exports = class List {
       return null;
     }
 
-    return this.itemQuery(context.authedItem.id, context, this.gqlNames.authenticatedQueryName);
+    return this.itemQuery(
+      { where: { id: context.authedItem.id } },
+      context,
+      this.gqlNames.authenticatedQueryName
+    );
   }
 
   get gqlMutationResolvers() {
@@ -743,18 +787,18 @@ module.exports = class List {
     }
 
     if (this.access.update) {
-      mutationResolvers[this.gqlNames.updateMutationName] = async (_, { id, data }, context) =>
+      mutationResolvers[this.gqlNames.updateMutationName] = (_, { id, data }, context) =>
         this.updateMutation(id, data, context);
 
-      mutationResolvers[this.gqlNames.updateManyMutationName] = async (_, { data }, context) =>
+      mutationResolvers[this.gqlNames.updateManyMutationName] = (_, { data }, context) =>
         this.updateManyMutation(data, context);
     }
 
     if (this.access.delete) {
-      mutationResolvers[this.gqlNames.deleteMutationName] = async (_, { id }, context) =>
+      mutationResolvers[this.gqlNames.deleteMutationName] = (_, { id }, context) =>
         this.deleteMutation(id, context);
 
-      mutationResolvers[this.gqlNames.deleteManyMutationName] = async (_, { ids }, context) =>
+      mutationResolvers[this.gqlNames.deleteManyMutationName] = (_, { ids }, context) =>
         this.deleteManyMutation(ids, context);
     }
 
@@ -776,8 +820,15 @@ module.exports = class List {
     });
   }
 
-  async _mapToFields(fields, action) {
-    return await resolveAllKeys(arrayToObject(fields, 'path', action));
+  _mapToFields(fields, action) {
+    return resolveAllKeys(arrayToObject(fields, 'path', action)).catch(error => {
+      if (!error.errors) {
+        throw error;
+      }
+      const errorCopy = new Error(error.message || error.toString());
+      errorCopy.errors = Object.values(error.errors);
+      throw errorCopy;
+    });
   }
 
   _fieldsFromObject(obj) {
@@ -813,7 +864,12 @@ module.exports = class List {
   }
 
   async _resolveInput(resolvedData, existingItem, context, operation, originalInput) {
-    const args = { resolvedData, existingItem, context, adapter: this.adapter, originalInput };
+    const args = {
+      resolvedData,
+      existingItem,
+      originalInput,
+      actions: mapKeys(this.hooksActions, hook => hook(context)),
+    };
     const fields = this._fieldsFromObject(resolvedData);
 
     resolvedData = await this._mapToFields(fields, field =>
@@ -834,15 +890,23 @@ module.exports = class List {
   }
 
   async _validateInput(resolvedData, existingItem, context, operation, originalInput) {
-    const args = { resolvedData, existingItem, context, adapter: this.adapter, originalInput };
+    const args = {
+      resolvedData,
+      existingItem,
+      originalInput,
+      actions: mapKeys(this.hooksActions, hook => hook(context)),
+    };
     const fields = this._fieldsFromObject(resolvedData);
-    this._validateHook(args, fields, operation, 'validateInput');
+    await this._validateHook(args, fields, operation, 'validateInput');
   }
 
   async _validateDelete(existingItem, context, operation) {
-    const args = { existingItem, context, adapter: this.adapter };
+    const args = {
+      existingItem,
+      actions: mapKeys(this.hooksActions, hook => hook(context)),
+    };
     const fields = this.fields;
-    this._validateHook(args, fields, operation, 'validateDelete');
+    await this._validateHook(args, fields, operation, 'validateDelete');
   }
 
   async _validateHook(args, fields, operation, hookName) {
@@ -873,23 +937,39 @@ module.exports = class List {
   }
 
   async _beforeChange(resolvedData, existingItem, context, originalInput) {
-    const args = { resolvedData, existingItem, context, adapter: this.adapter, originalInput };
-    this._runHook(args, resolvedData, 'beforeChange');
+    const args = {
+      resolvedData,
+      existingItem,
+      originalInput,
+      actions: mapKeys(this.hooksActions, hook => hook(context)),
+    };
+    await this._runHook(args, resolvedData, 'beforeChange');
   }
 
   async _beforeDelete(existingItem, context) {
-    const args = { existingItem, context, adapter: this.adapter };
-    this._runHook(args, existingItem, 'beforeDelete');
+    const args = {
+      existingItem,
+      actions: mapKeys(this.hooksActions, hook => hook(context)),
+    };
+    await this._runHook(args, existingItem, 'beforeDelete');
   }
 
   async _afterChange(updatedItem, existingItem, context, originalInput) {
-    const args = { updatedItem, originalInput, existingItem, context, adapter: this.adapter };
-    this._runHook(args, originalInput, 'afterChange');
+    const args = {
+      updatedItem,
+      originalInput,
+      existingItem,
+      actions: mapKeys(this.hooksActions, hook => hook(context)),
+    };
+    await this._runHook(args, updatedItem, 'afterChange');
   }
 
   async _afterDelete(existingItem, context) {
-    const args = { existingItem, context, adapter: this.adapter };
-    this._runHook(args, existingItem, 'afterDelete');
+    const args = {
+      existingItem,
+      actions: mapKeys(this.hooksActions, hook => hook(context)),
+    };
+    await this._runHook(args, existingItem, 'afterDelete');
   }
 
   async _runHook(args, fieldObject, hookName) {
@@ -906,7 +986,11 @@ module.exports = class List {
     // Set up a fresh mutation state if we're the root mutation
     const isRootMutation = !mutationState;
     if (isRootMutation) {
-      mutationState = { afterChangeStack: [], queues: {} };
+      mutationState = {
+        afterChangeStack: [], // post-hook stack
+        queues: {}, // backlink queues
+        transaction: {}, // transaction
+      };
     }
 
     // Perform the mutation
@@ -919,6 +1003,9 @@ module.exports = class List {
     const { afterChangeStack } = mutationState;
     afterChangeStack.push(afterHook);
     if (isRootMutation) {
+      // TODO: Close transation
+
+      // Execute post-hook stack
       while (afterChangeStack.length) {
         await afterChangeStack.pop()();
       }
@@ -936,7 +1023,9 @@ module.exports = class List {
 
     const existingItem = undefined;
 
-    this.checkFieldAccess(operation, existingItem, data, context, { gqlName });
+    const itemsToUpdate = [{ existingItem, data }];
+
+    this.checkFieldAccess(operation, itemsToUpdate, context, { gqlName });
 
     return await this._createSingle(data, existingItem, context, mutationState);
   }
@@ -947,11 +1036,11 @@ module.exports = class List {
 
     this.checkListAccess(context, operation, { gqlName });
 
-    const existingItem = undefined;
+    const itemsToUpdate = data.map(d => ({ existingItem: undefined, data: d }));
 
-    data.forEach(d => this.checkFieldAccess(operation, existingItem, d, context, { gqlName }));
+    this.checkFieldAccess(operation, itemsToUpdate, context, { gqlName });
 
-    return Promise.all(data.map(d => this._createSingle(d, existingItem, context, mutationState)));
+    return Promise.all(data.map(d => this._createSingle(d, undefined, context, mutationState)));
   }
 
   async _createSingle(data, existingItem, context, mutationState) {
@@ -1014,7 +1103,9 @@ module.exports = class List {
       gqlName,
     });
 
-    this.checkFieldAccess(operation, existingItem, data, context, { gqlName, extraData });
+    const itemsToUpdate = [{ existingItem, data }];
+
+    this.checkFieldAccess(operation, itemsToUpdate, context, { gqlName, extraData });
 
     return await this._updateSingle(id, data, existingItem, context, mutationState);
   }
@@ -1035,9 +1126,8 @@ module.exports = class List {
       data: data.map(d => d.data),
     });
 
-    itemsToUpdate.map(({ existingItem, data }) =>
-      this.checkFieldAccess(operation, existingItem, data, context, { gqlName, extraData })
-    );
+    // FIXME: We should do all of these in parallel and return *all* the field access violations
+    this.checkFieldAccess(operation, itemsToUpdate, context, { gqlName, extraData });
 
     return Promise.all(
       itemsToUpdate.map(({ existingItem, id, data }) =>