+++ packages/ui/src/primitives/navigation/pagination/Pagination.js
@@ -1,11 +1,15 @@
 // @flow
 
-import React, { Component } from 'react';
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
+import { Component } from 'react';
 import styled from '@emotion/styled';
 
 import { colors } from '../../../theme';
 import Page from './Page';
+import { LoadingSpinner } from '../../loading';
 import type { CountArgs, CountFormat, LabelType, OnChangeType } from './types';
+import { useEffect, useState } from '../../../new-typed-react';
 
 function ariaPageLabelFn(page: number) {
   return `Go to page ${page}`;
@@ -49,6 +53,7 @@ export type PaginationProps = {
   plural: string,
   singular: string,
   total: number,
+  isLoading: boolean,
 };
 
 const PaginationElement = styled.nav({
@@ -59,6 +64,31 @@ const PageCount = styled.div({
   color: colors.N60,
   marginRight: '1em',
 });
+
+const PageChildren = ({ page, isLoading, isSelected }) => {
+  const [shouldShowLoading, setShouldShowLoading] = useState(false);
+  useEffect(
+    () => {
+      if (isLoading && isSelected) {
+        const id = setTimeout(() => {
+          setShouldShowLoading(true);
+        }, 200);
+        return () => {
+          clearTimeout(id);
+          setShouldShowLoading(false);
+        };
+      }
+    },
+    [page, isLoading, isSelected]
+  );
+  return shouldShowLoading ? (
+    <div css={{ height: 19 }}>
+      <LoadingSpinner />
+    </div>
+  ) : (
+    <span>{page}</span>
+  );
+};
 
 class Pagination extends Component<PaginationProps> {
   static defaultProps = {
@@ -146,7 +176,7 @@ class Pagination extends Component<PaginationProps> {
           onClick={onChange}
           value={page}
         >
-          {page}
+          <PageChildren isLoading={this.props.isLoading} page={page} isSelected={isSelected} />
         </Page>
       );
     }