+++ website/src/pages/index.js
@@ -1,23 +1,46 @@
 import React from 'react';
 import { graphql, Link } from 'gatsby';
 
+import Search from '../components/search';
+
 export default ({ data }) => (
   <div>
+    <Search />
     <h1>
       Hello{' '}
       <span role="img" aria-label="Waving hand">
         👋
       </span>
     </h1>
+    <h2>
+      Start here{' '}
+      <span role="img" aria-label="hand pointing right">
+        👉
+      </span>{' '}
+      <Link to="/docs">/docs</Link>
+    </h2>
     {data.allSitePage.totalCount >= 1 ? (
       <div>
         <h2>Pages ({data.allSitePage.totalCount})</h2>
         <ul>
-          {data.allSitePage.edges.map(({ node }) => (
-            <li key={node.path}>
-              <Link to={node.path}>{node.path}</Link>
-            </li>
-          ))}
+          {data.allSitePage.edges
+            // Set up a particular order of results here:
+            // - '/' always comes first
+            // - '/docs' always comes next
+            // - ... the rest of the results, ordered by 'path' by the query
+            .sort((a, b) =>
+              a.node.path === '/' ||
+              (a.node.context.workspace !== b.node.context.workspace &&
+                a.node.context.workspace === 'docs' &&
+                b.node.path !== '/')
+                ? -1
+                : 0
+            )
+            .map(({ node }) => (
+              <li key={node.path}>
+                <Link to={node.path}>{node.path}</Link>
+              </li>
+            ))}
         </ul>
       </div>
     ) : (
@@ -33,6 +56,9 @@ export const pageQuery = graphql`
       edges {
         node {
           path
+          context {
+            workspace
+          }
         }
       }
     }